requirements.txt
fastapi
uvicorn
scikit-learn
pandas
python-whois
python-Levenshtein
dnspython
requests